% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlayCard.R
\name{overlayCard-shiny}
\alias{overlayCard-shiny}
\alias{overlayCardOutput}
\alias{renderOverlayCard}
\title{Shiny bindings for overlayCard}
\usage{
overlayCardOutput(outputId, width = "100\%", height = "400px")

renderOverlayCard(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a overlayCard}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using overlayCard within Shiny
applications and interactive Rmd documents.
}
